backend/requirements.txt
django==2.2.16
django-colorfield==0.7.2
django-cors-headers==3.11.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
Pillow==8.3.1
psycopg2-binary==2.8.6
reportlab==3.5